feat(sdk): history management
a History is essentially just a wrapper around a tuple of {key, counter}, which
can be used to manage changes.

the counter will be implemented when working with tasks, projects, and
checklists.

for now one can list all known histories, create new or delete existing onces.

As the keys are managed by the remote API there's no sense in exposing this to
users of this SDK